Decompose admin and user panels in components
